% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_pathway_enrich.R
\name{calculate_pvalue}
\alias{calculate_pvalue}
\title{This function is called by the \code{data_analysis} function.
calculate_pvalue}
\usage{
calculate_pvalue(data_extended, data_selected, genes_coverage, p_adjust_method)
}
\value{

}
\description{
It is used to calcualte the P value and the adjusted P value
for every Term per TAD.
}
