% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_enrich.R
\name{compare_density_plot}
\alias{compare_density_plot}
\title{This fuction is called from the \code{enrichr_visual} function.
It creates two density plots in the same graph.
compare_density_plot}
\usage{
compare_density_plot(string, type, data_visual)
}
\arguments{
\item{data_visual}{}
}
\value{

}
\description{

}
\details{
The first density plot is made with the P values of the data
after the analysis.The second density plot is made with the
Adjusted P values.
}
\examples{

}
