% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_pathway_enrich.R
\name{create_folders}
\alias{create_folders}
\title{This file contains functions used in the \verb{03_functional_analysis.R}
script.
This function is called by the \verb{03_functional_analysis.R} script.
create_folders}
\usage{
create_folders(output_folder)
}
\arguments{
\item{output_folder}{}
}
\value{

}
\description{
It creates the output folders.
}
\examples{

}
