% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_pathway_enrich.R
\name{data_analysis}
\alias{data_analysis}
\title{This function is called by the \code{functional_analysis.R} script.
data_analysis}
\usage{
data_analysis(
  enriched_terms,
  type,
  data_selected,
  genes_coverage,
  p_adjust_method,
  min_genes
)
}
\value{

}
\description{
It manipulates the enriched data and performs hypergeometric test.
}
