% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integration.R
\name{data_integration}
\alias{data_integration}
\title{Input parameters for Data Integration part}
\usage{
data_integration(
  dir_name = NULL,
  output_folder = NULL,
  tech = NULL,
  meta = NULL,
  counts_dir = NULL,
  freq_dir = NULL,
  tad_file = NULL
)
}
\arguments{
\item{dir_name}{Directory of input datasets containing feature counts and
frequency tables}

\item{output_folder}{Folder name for printing output tables}

\item{tech}{Human Genome Reference used}

\item{meta}{meta-data file name used}

\item{counts_dir}{Directory name of counts NGS data}

\item{freq_dir}{Directory name of freq NGS data}

\item{tad_file}{BED file containing information about TADs}
}
\value{

}
\description{

}
