% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_enrich.R
\name{density_plot}
\alias{density_plot}
\title{density_plot}
\usage{
density_plot(data_density, image_output_folder)
}
\value{

}
\description{
density_plot
}
\details{
This function is called by the \code{motif_visual} function.
It creates a density plot of the P values of the Motif EA.
}
\examples{

}
