% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_pathway_enrich.R
\name{enrich_per_tad}
\alias{enrich_per_tad}
\title{This function is called by the \verb{03_functional_analysis.R} script.
enrich_per_tad}
\usage{
enrich_per_tad(biodata, dbs, threshold, criterio, type)
}
\value{

}
\description{
It performs enrichment analysis using the Enrichr tool.
Enrichr input is the genes of the dataset grouped per TAD.
}
\examples{


}
