% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_enrich.R
\name{enrichr_visual}
\alias{enrichr_visual}
\title{This function is being called by the \verb{03_functional_analysis.R}
script.}
\usage{
enrichr_visual(folder, type, data_visual, criterio)
}
\arguments{
\item{criterio}{}
}
\value{

}
\description{

}
\details{
It creates graphs of the EA using Enrichr
enrichr_visual
}
\examples{

}
