% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evenDiff.R
\name{evenDiff}
\alias{evenDiff}
\title{evenDiff}
\usage{
evenDiff(
  mapping_file = NULL,
  methylo_result,
  names.meta = NULL,
  adj.PVal = NULL,
  log_thr = NULL
)
}
\arguments{
\item{mapping_file}{A meta-data file}

\item{methylo_result}{}

\item{names.meta}{meta data columns to process (names or indexes)}
}
\value{

}
\description{

}
\examples{
result<- data_integration (
counts_folder = system.file("extdata", "Datasets",
                         "counts", package="InterTADs"),
counts_fls = NULL,
freq_folder = system.file("extdata", "Datasets",
                         "freq", package="InterTADs"),
freq_fls = NULL,
mapping_file = system.file("extdata", "Datasets",
                         "meta-data.csv", package="InterTADs"),

tad_file =system.file("extdata", "Datasets",
                     "hglft_genome_2dab_ec1330.bed", package="InterTADs"),
tech = "hg38"
)

methylo_result <- prepare_methylation_values(
integratedTADtable = result[[1]],
mapping_file = system.file("extdata", "Datasets",
                         "meta-data.csv", package="InterTADs"),
meth_data = 2
)

result_evenDiff <- evenDiff(
mapping_file = system.file("extdata", "Datasets",
"meta-data.csv", package="InterTADs"),
methylo_result = methylo_result,
names.meta = c('group'),
adj.PVal = 0.01,
log_thr = 4)

print(result_evenDiff)


}
