% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_enrich.R
\name{group_plots}
\alias{group_plots}
\title{group_plots}
\usage{
group_plots(string, type, data_visual)
}
\arguments{
\item{data_visual}{}
}
\value{

}
\description{
The first folder is called "Adjusted P values per TADs".
Each histogram shows the Terms found in each TAD in ascending order,
according to their Adjusted P value.

The second folder is called "Adjusted P values per Term".
Each histogram shows the TADs, each Term was found into,
in ascending order, according to their Adjusted P value.


}
\details{
This function is called by the \code{enrichrVisual} function.
It creates two folders with histogram graphs.
}
\examples{

}
