% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_enrich.R
\name{histogram_plot}
\alias{histogram_plot}
\title{This function is called from the \code{enrichr_visual} function.
It creates two histogram graphs.}
\usage{
histogram_plot(string, type, data_visual)
}
\value{

}
\description{
The first one shows the Terms that were found in the most TADs
in descending order according to the number of TADs they were
found into.

The second one shows the same Terms, but adds the information of
the mean Adjusted P value of each Term to be found in these TADs.
The Terms are arranged in ascending order, according to the mean
Adjusted P value.The bar color corresponds to the number of TADs
each Term was found into.
}
\details{
histogram_plot
}
\examples{

}
