% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_entrez_ids.R
\name{map_entrez_ids}
\alias{map_entrez_ids}
\title{map_entrez_ids}
\usage{
map_entrez_ids(entrez.ids, tech = "hg19")
}
\arguments{
\item{tech}{Human Genome Reference used}
}
\value{

}
\description{
map_entrez_ids
}
