% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrich.R
\name{motif_outputs}
\alias{motif_outputs}
\title{This function is called by the \code{functional_analysis.R} script.
motif_outputs}
\usage{
motif_outputs(report_list)
}
\value{

}
\description{
It manipulates the enriched data after the analysis and creates
three output data.tables to be used for the Output csv files and
the visualization.
}
