% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_enrich.R
\name{motif_visual}
\alias{motif_visual}
\title{This function is being called by the \verb{03_functional_nalysis.R} script
motif_visual}
\usage{
motif_visual(
  image_output_folder,
  motif_output_folder,
  data_visual,
  report_list,
  criterio
)
}
\arguments{
\item{criterio}{}
}
\value{

}
\description{

}
\details{
It creates graphs of the Motif EA.
}
\examples{


}
