% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrich.R
\name{motifs_enrich}
\alias{motifs_enrich}
\title{This file contains functions used in the \verb{03_enrichmentAnalysis.R}
script.
This function is called by the \code{functional_analysis.R} script.
motifs_enrich}
\usage{
motifs_enrich(
  biodata,
  motif_output_folder,
  dir_name,
  p_adjust_method,
  cut_off,
  tech,
  exp_parent
)
}
\value{

}
\description{
It performs enrichment analysis using the PWMEnrich tool.
PWMEnrich input is the DNA sequences grouped per TAD.
}
