% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_enrich.R
\name{network_plot}
\alias{network_plot}
\title{network_plot}
\usage{
network_plot(string, type, data_visual)
}
\value{

}
\description{

}
\details{
This function is called from the \code{enrichr_visual} function.
It creates a network graph of the Terms (max 10),
that were found in the most TADs.

Each Term is a group.
Each node is unique and corresponds to a TAD.
The nodes are connected with lines.
Each different line color corresponds to a group.
}
\examples{

}
