% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_enrich.R
\name{per_tad_plots}
\alias{per_tad_plots}
\title{per_tad_plots}
\usage{
per_tad_plots(report_list, image_output_folder)
}
\value{

}
\description{
It creates a folder with graphs of the Motif EA for each TAD.
The folder is called "Plots per TADs".
For each TAD it creates a folder with two graphs.
The first one shows the summary of the Motif EA for this TAD,
it is called PWMEnrich_Image.
It is created with the \code{PWMEnrich::plot()} function.
The second one is called Histogram Image.
It shows the TFs found in the TAD, arranged in ascending order,
according to the mean P value.
It also returns the data.table data_density, with the P values
of the Motif EA of all TADs.
}
\details{
This function is called by the \code{motif_visual} function.
}
\examples{


}
