% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_enrich.R
\name{per_tfs_plots}
\alias{per_tfs_plots}
\title{per_tfs_plots}
\usage{
per_tfs_plots(report_list, data_visual, image_output_folder)
}
\value{

}
\description{
It creates a folder with graphs of the Motif EA for each TF found.
The folder is called "Transcription Factors".
For each TF it creates a folder with two graphs.
The first one shows the binding motifs of the TF
that were found in the analysis and is called Motifs.
It is created with the `ggseglogo::ggseqlogo()`` function.
The second one is called Barplot.
It shows the TADs that the TF was found into, arranged in
ascending order, according to the mean P value.
It also creates a histogram of the TFs found in most TADs,
called "Top Transcription Factors".
The TFs are arranged in descending order, according to the
number of TADs they were found into.
Lastly it creates a summary graph of the Top 10 TFs
and their Top binding Motifs.
It is called Top 10 Transcription Factors.
}
\details{
This function is called by the \code{motif_visual} function.
}
\examples{

}
