% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_methylation_values.R
\name{prepare_methylation_values}
\alias{prepare_methylation_values}
\title{Input parameters for TADiff part}
\usage{
prepare_methylation_values(integratedTADtable, mapping_file)
}
\arguments{
\item{dir_name}{Directory of input datasets containing feature counts
and frequency tables}

\item{output_folder}{Folder name for printing output tables}

\item{meta}{meta-data file name used}

\item{names.meta}{meta data columns to process (names or indexes)}

\item{meth_data}{Parent index of methylation data.
If no methylation is provided, place FALSE}
}
\value{

}
\description{

}
