% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrich.R
\name{prepare_sequences}
\alias{prepare_sequences}
\title{This function is called by the \code{motifs_enrich} function.
prepare_sequences}
\usage{
prepare_sequences(data, tech, motif_output_folder, exp_parent, dir_name)
}
\arguments{
\item{dir_name}{}
}
\value{

}
\description{
It is used to find the sequences, that correspond to TFBS
from the genomic coordinates of the events.
}
