% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_pathway_enrich.R
\name{produce_outputs}
\alias{produce_outputs}
\title{This function is called by the \code{data_analysis} function.
produce_outputs}
\usage{
produce_outputs(data_with_p, type)
}
\arguments{
\item{type}{the prevously selected databases acronyms used for the names
of the outputs files e.g. GO_MF for GO_Molecular_Function_2018}
}
\value{

}
\description{
It manipulates the enriched data after the analysis and creates
three output tables to be used for the Output csv files and the
visualization.
}
