% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_enrich.R
\name{set_graph_fonts}
\alias{set_graph_fonts}
\title{This file contains the functions used in the "Visualization"
section of the \verb{03_functional_analysis.R} script.
RStudio doesn't support all graph fonts in Windows OS.
This function sets the appropriate fonts for the graphs.}
\usage{
set_graph_fonts(device_system)
}
\arguments{
\item{device_system}{}
}
\value{

}
\description{

}
\details{
set_graph_fonts
}
\examples{

}
