% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integration.R
\name{data_integration}
\alias{data_integration}
\title{Input parameters for Data Integration part}
\usage{
data_integration(
  dir_name = "Datasets/",
  output_folder = "results_bloodcancer/",
  tech = "hg19",
  meta = "meta-data.csv",
  counts_dir = "counts",
  freq_dir = "freq",
  tad_file = "hglft_genome_2dab_ec1330.bed"
)
}
\arguments{
\item{dir_name}{Directory of input datasets containing feature counts and
frequency tables}

\item{output_folder}{Folder name for printing output tables}

\item{tech}{Human Genome Reference used}

\item{meta}{meta-data file name used}

\item{counts_dir}{Directory name of counts NGS data}

\item{freq_dir}{Directory name of freq NGS data}

\item{tad_file}{BED file containing information about TADs}
}
\value{

}
\description{

}
