% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integration.R
\name{data_integration}
\alias{data_integration}
\title{Input parameters for Data Integration part}
\usage{
data_integration(
  counts_folder = NULL,
  counts_fls = NULL,
  freq_folder = NULL,
  freq_fls = NULL,
  mapping_file = mapping_file,
  tad_file = tad_file,
  tech = "hg38"
)
}
\arguments{
\item{counts_folder}{Path to a repository of counts NGS data}

\item{counts_fls}{A vector of paths to files of counts NGS data}

\item{freq_folder}{Path to a repository of freq NGS data}

\item{freq_fls}{A vector of paths to files of freq NGS data}

\item{mapping_file}{A meta-data file}

\item{tad_file}{BED file containing information about TADs}

\item{tech}{Homo sapiens (human) genome assembly}
}
\value{

}
\description{

}
