% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integration.R
\name{data_integration}
\alias{data_integration}
\title{Data Integration part}
\usage{
data_integration(
  counts_folder = NULL,
  counts_fls = NULL,
  freq_folder = NULL,
  freq_fls = NULL,
  mapping_file,
  tad_file,
  tech
)
}
\arguments{
\item{counts_folder}{A stringr-path to a repository containing files
with NGS counts data (e.g. expression count table).}

\item{counts_fls}{A character vector of paths to files
with NGS counts data (e.g. expression count table).}

\item{freq_folder}{A stringr-path to a repository containing files
with NGS frequency data (e.g. DNA methylation table).}

\item{freq_fls}{A character vector of paths to files
with NGS frequency data (e.g. DNA methylation table).}

\item{mapping_file}{A mapping file containing mapping betwen the columns of the input
NGS datasets. The first column corresponds to the sample ID that
will be used in the output file. The following columns correspond
to the column names of the input datasets.
The file also contains the related sample meta-data}

\item{tad_file}{A BED file containing information
about the desired TADs' chromosomal location.}

\item{tech}{Human Genome Reference technology, either "hg19" or "hg38" can be used.}
}
\value{
A list of integratedTADtable and summary.
IntegratedTADtable contains all data info about TADs and
summary explains types of different events (methylation, expression etc.).
}
\description{
During the data integration parth different NGS datasets are
combined into the same data table alonng with the related input
data for the topologically associating domains (TADs).
}
