% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_pathway_enrich.R
\name{enrich_all}
\alias{enrich_all}
\title{This function is called by the \verb{03_functional_analysis.R} script.
enrich_all}
\usage{
enrich_all(biodata, dbs, threshold, criterio, type)
}
\value{

}
\description{
It performs enrichment analysis using the Enrichr tool.
Enrichr input is all the genes of the dataset.
}
\examples{

}
