% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_ids.R
\name{ensembl_ids}
\alias{ensembl_ids}
\title{ensembl_ids}
\usage{
ensembl_ids(input_file, expr_data)
}
\arguments{
\item{input_file}{}

\item{expr_data}{Parent index of expression data.}
}
\value{
A table with matched ensembl_ids
}
\description{

}
\examples{
result<- data_integration (
counts_folder = system.file("extdata", "Datasets",
                         "counts", package="InterTADs"),
counts_fls = NULL,
freq_folder = system.file("extdata", "Datasets",
                         "freq", package="InterTADs"),
freq_fls = NULL,
mapping_file = system.file("extdata", "Datasets",
                         "meta-data.csv", package="InterTADs"),

tad_file =system.file("extdata", "Datasets",
                     "hglft_genome_2dab_ec1330.bed", package="InterTADs"),
tech = "hg38"
)
methylo_result <- prepare_methylation_values(
integratedTADtable = result[[1]],
mapping_file = system.file("extdata", "Datasets",
                         "meta-data.csv", package="InterTADs"),
meth_data = 2
)

result_ensmbl <- ensembl_ids(
input_file= methylo_result,
expr_data = 3)


}
