% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_ids.R
\name{ensembl_ids}
\alias{ensembl_ids}
\title{ensembl_ids}
\usage{
ensembl_ids(dir_name = NULL, output_folder = NULL, expr_data = NULL)
}
\arguments{
\item{dir_name}{Directory of input datasets containing feature counts and
frequency tables}

\item{output_folder}{Folder name for printing output tables}

\item{expr_data}{Parent index of expression data.
If no expression is provided, place FALSE}
}
\value{
TRUE if function runs properly
}
\description{

}
\examples{
result_ensmbl <- ensembl_ids(dir_name = "Datasets",
output_folder = 'results_bloodcancer',
expr_data = 2)
print(result_ensmbl)

}
