% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evenDiff.R
\name{evenDiff}
\alias{evenDiff}
\title{evenDiff}
\usage{
evenDiff(
  dir_name = NULL,
  input_f = NULL,
  output_folder = NULL,
  meta = NULL,
  names.meta = NULL
)
}
\arguments{
\item{dir_name}{Directory of input datasets containing feature counts and
frequency tables}

\item{output_folder}{Folder name for printing output tables}

\item{meta}{meta-data file name used}

\item{names.meta}{meta data columns to process (names or indexes)}
}
\value{
TRUE if function runs properly
}
\description{

}
\examples{
result_evenDiff <- evenDiff(
dir_name = system.file("extdata","Datasets",package='InterTADs'),
input_f = system.file("extdata","results_bloodcancer",
"integrated-tad-table-methNorm.txt",package='InterTADs'),
output_folder = system.file("extdata","results_bloodcancer",
package='InterTADs'),
meta = "meta-data.csv",
names.meta = c('groups'))

print(result_evenDiff)

}
