% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evenDiff.R
\name{evenDiff}
\alias{evenDiff}
\title{evenDiff}
\usage{
evenDiff(dir_name = NULL, output_folder = NULL, meta = NULL, names.meta = NULL)
}
\arguments{
\item{dir_name}{Directory of input datasets containing feature counts and
frequency tables}

\item{output_folder}{Folder name for printing output tables}

\item{meta}{meta-data file name used}

\item{names.meta}{meta data columns to process (names or indexes)}
}
\value{
TRUE if function runs properly
}
\description{

}
\examples{
result_evenDiff <- evenDiff(dir_name = "Datasets",
output_folder = "results_bloodcancer",
meta = "meta-data.csv",
names.meta = c('groups'))

print(result_evenDiff)


}
