% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis.R
\name{functional_analysis}
\alias{functional_analysis}
\title{Functional analysis.}
\usage{
functional_analysis(
  tech = "hg19",
  exp_parent = 2,
  dbs = c("GO_Molecular_Function_2018", "GO_Biological_Process_2018",
    "KEGG_2019_Human"),
  type = c("GO_MF", "GO_BP", "KEGG"),
  genes_cover = c(11459, 14433, 7802),
  p_adjust_method = "fdr",
  cut_off = 0.05,
  criterio = "adjusted p-value",
  min_genes = 3,
  system = "win",
  dir_name = "Datasets"
)
}
\arguments{
\item{tech}{Human Genome Reference used}

\item{exp_parent}{number of the parent file of the expression data}

\item{dbs}{Databases used from EnrichR}

\item{type}{the prevously selected databases acronyms used for the names
of the outputs files e.g. GO_MF for GO_Molecular_Function_2018}

\item{genes_cover}{gene coverage of the databases}

\item{p_adjust_method}{p adjustment method, the methods supported are:
\verb{c("holm", "hochberg", "hommel", "bonferroni",}
\verb{"BH", "BY", "fdr", "none")}}

\item{cut_off}{cut-off Enrichr enrichment (adjusted) p-value}

\item{criterio}{Enrichr result column selected as criterio:
\code{"p-value"} or \code{"adjusted p-value"}}

\item{min_genes}{min number of genes in over-represented terms}

\item{system}{RStudio supports different fonts for different
operating systems}

\item{dir_name}{name or filepath of the input folder}

\item{output_folder}{name or filepath of the output folder}
}
\value{

}
\description{
Enrichment analysis with GO terms, KEGG pathways and motif enrichment
with TFs.
Optional step to be performed on the output files
of the \verb{02c_evenDiff.R} and \verb{02d_TADiff.R} scripts.


}
