% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrich.R
\name{get_dna_sequences}
\alias{get_dna_sequences}
\title{This function is called by the \code{motifs_enrich} function.
get_dna_sequences}
\usage{
get_dna_sequences(input_data, outputs_folder, tech)
}
\arguments{
\item{tech}{}
}
\value{

}
\description{
It is used to query the Rest Ensembl API.
It gets the DNA sequences that correspond to the genomic
coordinates of the events.
}
