% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_methylation_values.R
\name{prepare_methylation_values}
\alias{prepare_methylation_values}
\title{prepare_methylation_values}
\usage{
prepare_methylation_values(integratedTADtable, mapping_file, meth_data)
}
\arguments{
\item{integratedTADtable}{IntegratedTADtable contains
all data info about TADs}

\item{mapping_file}{A meta-data file}

\item{meth_data}{Parent index of methylation data, from meta-data.csv.
If no methylation is provided, place FALSE}
}
\value{

}
\description{

}
\examples{

result<- data_integration (
counts_folder = system.file("extdata", "Datasets",
                         "counts", package="InterTADs"),
counts_fls = NULL,
freq_folder = system.file("extdata", "Datasets",
                         "freq", package="InterTADs"),
freq_fls = NULL,
mapping_file = system.file("extdata", "Datasets",
                         "meta-data.csv", package="InterTADs"),

tad_file =system.file("extdata", "Datasets",
                     "hglft_genome_2dab_ec1330.bed", package="InterTADs"),
tech = "hg38"
)
methylo_result <- prepare_methylation_values(
integratedTADtable = result[[1]],
mapping_file = system.file("extdata", "Datasets",
                         "meta-data.csv", package="InterTADs"),
meth_data = 2
)
}
