% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrich.R
\name{prepare_sequences}
\alias{prepare_sequences}
\title{This function is called by the \code{motifs_enrich} function.
prepare_sequences}
\usage{
prepare_sequences(
  data,
  tech,
  output_folder,
  exp_parent,
  dir_name,
  annotation_file
)
}
\arguments{
\item{tech}{Human Genome Reference used}

\item{output_folder}{name or filepath of the output folder}

\item{exp_parent}{number of the parent file of the expression data}

\item{dir_name}{name or filepath of the input folder}
}
\value{

}
\description{
It is used to find the sequences, that correspond to TFBS
from the genomic coordinates of the events.
}
