% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TADiff.R
\name{TADiff}
\alias{TADiff}
\title{Input parameters for TADiff part}
\usage{
TADiff(
  dir_name = NULL,
  output_folder = NULL,
  meta = NULL,
  names.meta = NULL,
  expr_data = NULL
)
}
\arguments{
\item{dir_name}{Directory of input datasets containing feature counts and
frequency tables}

\item{output_folder}{Folder name for printing output tables}

\item{meta}{meta-data file name used}

\item{names.meta}{meta data columns to process (names or indexes)}

\item{expr_data}{Parent index of expression data.
If no expression is provided, place FALSE}
}
\value{

}
\description{

}
