% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TADiff.R
\name{TADiff}
\alias{TADiff}
\title{TADiff}
\usage{
TADiff(
  integratedTADtable = NULL,
  mapping_file = NULL,
  names.meta = NULL,
  expr_data = NULL,
  adj.PVal = 0.05,
  log_thr = 2,
  tad_event = 4,
  pval_thr = 0.05,
  freq_thr = 15,
  mean_logFC_thr = 2,
  seed = 6
)
}
\arguments{
\item{integratedTADtable}{IntegratedTADtable contains all data info about TADs.
This is the output integrated table
from data_integration, prepare_methylation or
exnsembl_ids functions.}

\item{mapping_file}{A mapping file containing mapping betwen the columns of the input
NGS datasets. The first column corresponds to the sample ID that
will be used in the output file. The following columns correspond
to the column names of the input datasets.
The file also contains the related sample meta-data}

\item{names.meta}{meta data columns to process (names or indexes)}

\item{expr_data}{Parent index of expression data.}

\item{adj.PVal}{Significant events adjusted p value theshold. Defaults to 0.05.}

\item{log_thr}{LogFC theshold. Defaults to 2}

\item{tad_event}{Number of events for each TAD. Defaults to 4.}

\item{pval_thr}{Significant TADs p value threshold. Defaults to 0.05.}

\item{freq_thr}{Percentage theshold of significant events to all events for each TAD, Defaults to 20 (20\%).}

\item{mean_logFC_thr}{Mean logFC threshold for each TAD.}

\item{seed}{A number used to initialize a pseudorandom number generator.}
}
\value{
A list of data tables containing significant TADs along with
a summary file
}
\description{
TADiff performs differential analysis of events (expression, methylation etc.)
by taking into account the chromatin configuration of the genome,
i.e. the topologically associating domains (TADs).
}
\examples{

result <- data_integration (

    counts_folder = system.file(
        "extdata", "Datasets", "counts", package = "InterTADs"
    ),

    freq_folder = system.file(
        "extdata", "Datasets", "freq", package = "InterTADs"
    ),

    mapping_file = system.file(
       "extdata", "Datasets", "meta-data.csv", package = "InterTADs"
   ),

    tad_file =system.file(
       "extdata", "Datasets",
       "hglft_genome_2dab_ec1330.bed", package = "InterTADs"
    ),

    tech = "hg19"
)


methylo_result <- prepare_methylation_values(
    integratedTADtable = result[[1]],
    mapping_file = system.file(
        "extdata", "Datasets", "meta-data.csv", package = "InterTADs"
    ),
    meth_data = 2
)

result_ensmbl <- ensembl_ids(
    integratedTADtable = methylo_result,
    expr_data = 3
)

TADiff_result <- TADiff(

    integratedTADtable = result_ensmbl,
    mapping_file = system.file(
        "extdata", "Datasets", "meta-data.csv", package = "InterTADs"
    ),
    names.meta = c("group"),
    expr_data = 3,
    adj.PVal = 0.01,
    log_thr = 2,
    tad_event = 4,
    pval_thr = 0.01,
    freq_thr = 20,
    mean_logFC_thr = 2
)
}
